Send block header to asic request task
* Move block header construction out of asic component and into job component
* Update `send_work` to append the work ID & and num_midstates "properties"
* asic request task queue now just expects 80 bytes
